Contributor controller
